Domain aggregate rename
