--- src/core/Codend.Domain/Core/Abstractions/IAggregate.cs
@@ -1,8 +1,0 @@
-namespace Codend.Domain.Core.Abstractions;
-
-public interface IAggregate
-{
-    IReadOnlyCollection<IDomainEvent> DomainEvents { get; }
-
-    void ClearDomainEvents();
-}
