--- src/core/Codend.Domain/Core/Events/ProjectTask/ProjectTaskAddedToProjectEvent.cs
@@ -1,19 +1,0 @@
-﻿using Codend.Domain.Core.Abstractions;
-using Codend.Domain.Entities;
-
-namespace Codend.Domain.Core.Events;
-
-/// <summary>
-/// Domain event raised after adding new task to project.
-/// </summary>
-public class ProjectTaskAddedToProjectEvent : IDomainEvent
-{
-    public ProjectTaskAddedToProjectEvent(BaseProjectTask projectTask, ProjectId projectId)
-    {
-        ProjectTask = projectTask;
-        ProjectId = projectId;
-    }
-
-    public BaseProjectTask ProjectTask { get; set; }
-    public ProjectId ProjectId { get; set; }
-}
