--- src/core/Codend.Domain/Core/Events/ProjectTask/ProjectTaskDeletedFromProjectEvent.cs
@@ -1,19 +1,0 @@
-﻿using Codend.Domain.Core.Abstractions;
-using Codend.Domain.Entities;
-
-namespace Codend.Domain.Core.Events;
-
-/// <summary>
-/// Domain event raised after deleting task from project.
-/// </summary>
-public class ProjectTaskDeletedFromProjectEvent : IDomainEvent
-{
-    public ProjectTaskDeletedFromProjectEvent(BaseProjectTask projectTask, ProjectId projectId)
-    {
-        ProjectTask = projectTask;
-        ProjectId = projectId;
-    }
-
-    public BaseProjectTask ProjectTask { get; set; }
-    public ProjectId ProjectId { get; set; }
-}
