--- src/core/Codend.Domain/Core/Events/ProjectTask/ProjectTaskNameEditedEvent.cs
@@ -1,20 +1,0 @@
-﻿using Codend.Domain.Core.Abstractions;
-using Codend.Domain.Entities;
-using Codend.Domain.ValueObjects;
-
-namespace Codend.Domain.Core.Events;
-
-/// <summary>
-/// Domain event raised after ProjectTaskName has been changed
-/// </summary>
-public class ProjectTaskNameEditedEvent : IDomainEvent
-{
-    public ProjectTaskNameEditedEvent(ProjectTaskName name, ProjectTaskId projectTaskId)
-    {
-        Name = name;
-        ProjectTaskId = projectTaskId;
-    }
-
-    public ProjectTaskName Name { get; set; }
-    public ProjectTaskId ProjectTaskId { get; set; }
-}
