--- src/core/Codend.Domain/Core/Primitives/Aggregate.cs
@@ -1,24 +1,0 @@
-﻿using Codend.Domain.Core.Abstractions;
-
-namespace Codend.Domain.Core.Primitives;
-
-public abstract class Aggregate<TKey> : Entity<TKey>, IAggregate
-{
-    protected Aggregate(TKey id) : base(id)
-    {
-    }
-
-    private readonly List<IDomainEvent> _domainEvents = new();
-
-    public IReadOnlyCollection<IDomainEvent> DomainEvents => this._domainEvents;
-
-    public void ClearDomainEvents()
-    {
-        _domainEvents.Clear();
-    }
-
-    protected void Raise(IDomainEvent domainEvent)
-    {
-        this._domainEvents.Add(domainEvent);
-    }
-}
